backend/requirements.txt
asgiref==3.3.2
Django==3.2.7
djangorestframework
markdown
django-filter
django-colorfield
pillow
djangorestframework-simplejwt
djoser
reportlab
pytest
flake8
gunicorn==20.0.4
psycopg2-binary
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
drf-extra-fields